fix: Move onClick handler to a component
